+++ commons/src/test/java/io/hyscale/commons/utils/HyscaleStringUtilsTests.java
@@ -30,6 +30,11 @@ public class HyscaleStringUtilsTests {
     private static final char TAILING_CHAR = 'r';
     private static final String EXPECTED_CHAR_STRING = "hyscaleUse";
 
+    private static String[] singularWords = {"policy", "resource", "scarf", "knife", "mango", "backslash",
+            "video", "outlay", "bus", "fix", "buzz", "beach"};
+    private static String[] pluralWords = {"policies", "resources", "scarves", "knives", "mangoes", "backslashes",
+            "videos", "outlays", "buses", "fixes", "buzzes", "beaches"};
+
     public static Stream<Arguments> getInputsForRemoveSuffixTest() {
         return Stream.of(Arguments.of(null, SUFFIX_STRING, null),
                 Arguments.of("", SUFFIX_STRING, ""),
@@ -37,6 +42,29 @@ public class HyscaleStringUtilsTests {
                 Arguments.of(SAMPLE_STRING, "", SAMPLE_STRING),
                 Arguments.of(SAMPLE_STRING, SUFFIX_STRING, EXPECTED_STRING));
     }
+
+    public static Stream<Arguments> singularToPluralTest() {
+        return Stream.of(Arguments.of(singularWords[0], pluralWords[0]),
+                Arguments.of(singularWords[1], pluralWords[1]),
+                Arguments.of(singularWords[2], pluralWords[2]),
+                Arguments.of(singularWords[3], pluralWords[3]),
+                Arguments.of(singularWords[4], pluralWords[4]),
+                Arguments.of(singularWords[5], pluralWords[5]),
+                Arguments.of(singularWords[6], pluralWords[6]),
+                Arguments.of(singularWords[7], pluralWords[7]),
+                Arguments.of(singularWords[8], pluralWords[8]),
+                Arguments.of(singularWords[9], pluralWords[9]),
+                Arguments.of(singularWords[10], pluralWords[10]),
+                Arguments.of(singularWords[11], pluralWords[11]));
+    }
+
+    @ParameterizedTest
+    @MethodSource(value = "singularToPluralTest")
+    void testPluralFunction(String singular, String plural) {
+        String output = HyscaleStringUtil.getPlural(singular);
+        assertEquals(plural, output);
+    }
+
 
     @ParameterizedTest
     @MethodSource(value = "getInputsForRemoveSuffixTest")