+++ deployer/deployer-core/src/main/java/io/hyscale/deployer/core/model/ResourceKind.java
@@ -33,6 +33,7 @@ public enum ResourceKind {
     EVENT("Event"),
     VERSION("VersionInfo"),
     INGRESS("Ingress");
+    NETWORKPOLICY("NetworkPolicy");
 
     private String kind;