+++ deployer/deployer-services/src/main/java/io/hyscale/deployer/services/handler/ResourceLifeCycleHandler.java
@@ -236,4 +236,7 @@ public interface ResourceLifeCycleHandler<T> {
         return ResourceUpdatePolicy.PATCH;
     }
 
+    default boolean isWorkLoad() {
+    	return false;
+    }
 }