+++ deployer/deployer-services/src/main/java/io/hyscale/deployer/services/processor/PodParentUtil.java
@@ -20,9 +20,17 @@ import com.google.gson.JsonParser;
 import io.hyscale.commons.models.AnnotationKey;
 import io.hyscale.commons.models.LoadBalancer;
 import io.hyscale.commons.utils.GsonProviderUtil;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import io.hyscale.commons.models.AnnotationKey;
+import io.hyscale.commons.utils.HyscaleContextUtil;
 import io.hyscale.commons.utils.ResourceSelectorUtil;
 import io.hyscale.deployer.core.model.CustomResourceKind;
-import io.hyscale.deployer.core.model.ResourceKind;
 import io.hyscale.deployer.services.client.GenericK8sClient;
 import io.hyscale.deployer.services.client.K8sResourceClient;
 import io.hyscale.deployer.services.model.CustomObject;
@@ -34,8 +42,6 @@ import io.kubernetes.client.openapi.models.V1StatefulSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
-
 public class PodParentUtil {
 
     private static final String podParentApiVersion = "apps/v1";
@@ -44,18 +50,15 @@ public class PodParentUtil {
 
     private ApiClient apiClient;
     private String namespace;
-    private List<CustomResourceKind> workloadResources;
 
     public PodParentUtil(ApiClient apiClient, String namespace){
         this.apiClient = apiClient;
         this.namespace = namespace;
-        this.workloadResources = new ArrayList<>();
-        this.workloadResources.add(new CustomResourceKind(ResourceKind.DEPLOYMENT.getKind(), podParentApiVersion));
-        this.workloadResources.add(new CustomResourceKind(ResourceKind.STATEFUL_SET.getKind(), podParentApiVersion));
+        
     }
 
     public PodParent getPodParentForService(String serviceName){
-        for(CustomResourceKind customResourceKind : workloadResources){
+        for(CustomResourceKind customResourceKind : HyscaleContextUtil.getSpringBean(WorkloadKinds.class).get()) {
             GenericK8sClient genericK8sClient = new K8sResourceClient(apiClient).
                     withNamespace(namespace).forKind(customResourceKind);
             CustomObject resource = genericK8sClient.getResourceByName(serviceName);
@@ -71,16 +74,13 @@ public class PodParentUtil {
         Map<String,PodParent> serviceVsPodParents = new HashMap<>();
         GenericK8sClient genericK8sClient = null;
         List<CustomObject> workloadResources = new ArrayList<>();
-        genericK8sClient = new K8sResourceClient(apiClient).
-                withNamespace(namespace).forKind(new CustomResourceKind(ResourceKind.DEPLOYMENT.getKind(),podParentApiVersion));
-        //Fetch Deployments
-        workloadResources.addAll(genericK8sClient.getBySelector(selector));
-
-        genericK8sClient = new K8sResourceClient(apiClient).
-                withNamespace(namespace).forKind(new CustomResourceKind(ResourceKind.STATEFUL_SET.getKind(),podParentApiVersion));
-        // Fetch StatefulSets
-        workloadResources.addAll(genericK8sClient.getBySelector(selector));
-
+        
+		for (CustomResourceKind customResourceKind : HyscaleContextUtil.getSpringBean(WorkloadKinds.class).get()) {
+			genericK8sClient = new K8sResourceClient(apiClient).withNamespace(namespace).forKind(customResourceKind);
+			// Fetch Deployments
+			workloadResources.addAll(genericK8sClient.getBySelector(selector));
+		}
+        
         if(workloadResources != null && !workloadResources.isEmpty()){
             for(CustomObject customObject :workloadResources){
                 if(customObject.getMetadata() != null){