+++ manifest-generator/generator-services/src/main/java/io/hyscale/generator/services/provider/PluginTemplateProvider.java
@@ -19,13 +19,8 @@ import io.hyscale.commons.models.ConfigTemplate;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
-import java.io.File;
-import java.net.URL;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 @Component
 public class PluginTemplateProvider {
@@ -68,6 +63,8 @@ public class PluginTemplateProvider {
 
         ISTIO_GATEWAY("/loadBalancer/istio/gateway.yaml.tpl");
 
+        NETWORK_POLICY("networkPolicy.yaml.tpl");
+    		
         String templateFile;
 
         PluginTemplateType(String tplFile) {