+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/activity/ValidatorActivity.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2019 Pramati Prism, Inc.
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -30,7 +30,7 @@ public enum ValidatorActivity implements Activity {
     VALIDATING_MANIFEST("Validating manifest for service {} "),
     VALIDATING_REGISTRY("Validating registry {} "),
     VALIDATING_VOLUME("Validating volumes for service {} "),
-    
+
     VOLUME_VALIDATION_FAILED("Volume validation failed. Error {}"),
     MANIFEST_VALIDATION_FAILED("Manifest validation failed. Error {}"),
     DOCKER_NOT_INSTALLED("Docker not installed. Install docker to continue "),
@@ -39,7 +39,10 @@ public enum ValidatorActivity implements Activity {
     SERVICE_SPEC_VALIDATION_FAILED("Service spec validation failed for {}. {} "),
     PROFILE_VALIDATION_FAILED("Profile validation failed for {}. {} "),
     CLUSTER_AUTHENTICATION_FAILED("Cluster authentication failed. Verify if cluster config is valid"),
-
+    MISSING_PORTS("Invalid Service Spec. Missing required field 'ports' in allowTraffic."),
+    INVALID_EXTERNAL_VALUE("Invalid Service Spec. Traffic rules cannot be applied to a service exposed externally."),
+    DUPLICATE_PORTS("Invalid Service Spec. Duplicate ports found : {}"),
+    PORT_NOT_EXPOSED("Invalid Service Spec. Traffic rules cannot be applied to a port that is not exposed on the Service");
     //Load balancer validation
     PORTS_MISMATCH("Ports declared in the load balancer should match with the ports mentioned in the hspec. Mismatched port : {}"),
     EXTERNAL_CONFIGURED("When external is configured as false in hspec, loadBalancer configuration snippet can not be applied."),//TODO Verify warning message