+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/commands/deploy/HyscaleDeployServiceCommand.java
@@ -160,6 +160,11 @@ public class HyscaleDeployServiceCommand implements Callable<Integer> {
     @Autowired
     private LoadBalancerValidator loadBalancerValidator;
 
+    private PortsValidator portsValidator;
+
+    @Autowired
+    private NetworkPoliciesValidator networkPoliciesValidator;
+
     private List<Validator<WorkflowContext>> postValidators;
 
     private JsonArray jsonArr;
@@ -175,6 +180,8 @@ public class HyscaleDeployServiceCommand implements Callable<Integer> {
         this.postValidators.add(clusterValidator);
         this.postValidators.add(volumeValidator);
         this.postValidators.add(loadBalancerValidator);
+        this.postValidators.add(portsValidator);
+        this.postValidators.add(networkPoliciesValidator);
         this.jsonArr = new JsonArray();
         this.jsonParser = new JsonParser();
     }